rest/requirements.txt
Pillow
flask-cors
flask